html/editing/dnd/the-drag-data-store/DataTransferItem_getAsString.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>DataTransferItem Test: getAsString()</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<meta name="flags" content="interact">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<p><input id="button" type="text" value="dragcharacters" style="border:2px blue solid; width:200px; height: 100px;"/></p>
<p><input id="div" type="text" style="border:2px green solid; width:200px; height: 100px;"/></p>

<p>Select all characters in blue box and drag to green box then drop on the green box</p>

<script>

var div = document.getElementById("div");

setup({explicit_done: true});
setup({explicit_timeout: true});

on_event(div, "drop", function (evt) {
  var item = evt.dataTransfer.items[0];

  test(function () {
    var file1 = item.getAsFile();
    assert_equals(file1, null);
  }, "Check if DataTransferItem.getAsFile return null if drag data item kind is not File");

  test(function () {
    item.getAsString(function (data) {
      assert_equals(data, "dragcharacters");
      done();
    });
   }, "Check if DataTransferItem.getAsString return the dragged string");
});

</script>

<div id="log"></div>
